pyproject.toml
[project]
name = "yt-live-recorder"
version = "0.1.0"
description = "YouTube Live Recorder - A command-line tool for recording YouTube live streams"
requires-python = ">=3.9"
dependencies = [
    "yt-dlp>=2024.1.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
]

[project.scripts]
yt-recorder = "src.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E501"]
